PROMIS Anxiety Scale Dataset

The data contains responses from 766 people sampled from a general population to the PROMIS Anxiety scale (http://www.nihpromis.org) composed of 29 Likert-type questions with a common rating scale (1 = Never, 2 = Rarely, 3 = Sometimes, 4 = Often, and 5 = Always).

Usage

Anxiety

Format

A data frame with 766 observations on the following 32 variables.

age

0 = younger than 65 and 1 = 65 and older

gender

0 = Male and 1 = Female

education

0 = some college or higher and 1 = high school or lower

R1

I felt fearful

R2

I felt frightened

R3

It scared me when I felt nervous

R4

I felt anxious

R5

I felt like I needed help for my anxiety

R6

I was concerned about my mental health

R7

I felt upset

R8

I had a racing or pounding heart

R9

I was anxious if my normal routine was disturbed

R10

I had sudden feelings of panic

R11

I was easily startled

R12

I had trouble paying attention

R13

I avoided public places or activities

R14

I felt fidgety

R15

I felt something awful would happen

R16

I felt worried

R17

I felt terrified

R18

I worried about other people's reactions to me

R19

I found it hard to focus on anything other than my anxiety

R20

My worries overwhelmed me

R21

I had twitching or trembling muscles

R22

I felt nervous

R23

I felt indecisive

R24

Many situations made me worry

R25

I had difficulty sleeping

R26

I had trouble relaxing

R27

I felt uneasy

R28

I felt tense

R29

I had difficulty calming down

score

Total score.

zscore

Standardized total score.

Source

Reexport from lordif package; http://www.nihpromis.org

References

PROMIS Cooperative Group. Unpublished Manual for the Patient-Reported Outcomes Measurement Information System (PROMIS) Version 1.1. October, 2008: http://www.nihpromis.org