Homeostasis concept inventory dichotomous dataset

(HCI) dataset consists of the dichotomously scored responses of 651 students (405 males, 246 females) to Homeostasis Concept Inventory (HCI) multiple-choice test. It containts 20 items, vector of gender membership and identificator whether students plan to major.

data(HCI)

Format

HCI is a data.frame consisting of 651 observations on the 22 variables.

Item1-Item20

Dichotomously scored items of the HCI test.

gender

Gender membership, "0" males, "1" females.

major

Identificator whether student plans to major in the life sciences.

References

McFarland, J. L., Price, R. M., Wenderoth, M. P., Martinkova, P., Cliff, W., Michael, J., ... & Wright, A. (2017). Development and validation of the homeostasis concept inventory. CBE-Life Sciences Education, 16(2), ar35. doi: 10.1187/cbe.16-10-0305

See also

Author

Jenny L. McFarland
Biology Department, Edmonds Community College