Clinical outcomes in multiple sclerosis patients dataset

The MSclinical dataset contains clinical measures on multiple sclerosis patients.

data(MSclinical)

Format

MSclinical is a data.frame consisting of 17 observations on 13 variables.

LCLA

Low-Contrast Letter Acuity test.

MI

Motricity Index.

MAS

Modified Ashworth Scale.

BBS

Berg Balance Scale.

T

Tremor.

DD

Dysdiadochokinesia.

DM

Dysmetria.

PRs

Postural reactions.

KH

Knee Hyperextension.

NHPT

Nine-Hole Peg Test.

T25FW

Timed 25-Foot Walk.

PASAT3

3-minute version of the Paced Auditory Serial Addition Test.

EDSS

Kurtzke Expanded Disability Status Scale.

References

Rasova, K., Martinkova, P., Vyskotova, J., & Sedova, M. (2012). Assessment set for evaluation of clinical outcomes in multiple sclerosis: Psychometric properties. Patient related outcome measures, 3, 59. doi: 10.2147/PROM.S32241