Obtain model definition for mirt's nominal model taking in account the key of correct answers

Standard mirt model with itemtype = "nominal" puts the identification constrains on the item response category slopes such as \(ak_0 = 0\) and \(ak_{(K-1)} = (K - 1)\), freely estimating the rest.

While nominal item responses are unordered by definition, it is often the case that one of the item response categories is correct and the respondents endorsing this category "naturally" possess a higher latent ability. Use this function to obtain model definition where the correct response category \(k_c\) for item \(i\) with \(K\) possible response categories translates to constrains \(ak_{k_c} = (K - 1)\) and \(ak_{k_{d1}} = 0\), with \(k_{d1}\) being the first incorrect response category (i.e. the first distractor).

Usage

obtain_nrm_def(data_with_key, ...)

Arguments

data_with_key

The output of nominal_to_int().

...

arguments passed onto mirt::mirt(). No practical use for now.

Value

A data.frame with the starting values, parameter numbers, estimation constrains etc. Pass it as pars argument of mirt::mirt().

See also

Examples

library(mirt)

# convert nominal data to integers and the original labels with correct answers
data_with_key <- nominal_to_int(HCItest[, 1:20], HCIkey)

# build model definition for {mirt} using the returned list from above
nrm_def <- obtain_nrm_def(data_with_key)

# fit the nominal model using the obtained model definition in `pars` argument
fit <- mirt(data_with_key$Data, 1, "nominal", pars = nrm_def)
#> Iteration: 1, Log-Lik: -13961.342, Max-Change: 7.44447Iteration: 2, Log-Lik: -12029.045, Max-Change: 2.28698Iteration: 3, Log-Lik: -11785.296, Max-Change: 2.21681Iteration: 4, Log-Lik: -11700.210, Max-Change: 0.34664Iteration: 5, Log-Lik: -11659.239, Max-Change: 0.33643Iteration: 6, Log-Lik: -11625.410, Max-Change: 0.35871Iteration: 7, Log-Lik: -11586.678, Max-Change: 0.27133Iteration: 8, Log-Lik: -11563.789, Max-Change: 0.26793Iteration: 9, Log-Lik: -11540.559, Max-Change: 0.25421Iteration: 10, Log-Lik: -11521.396, Max-Change: 0.27580Iteration: 11, Log-Lik: -11505.697, Max-Change: 0.26654Iteration: 12, Log-Lik: -11493.529, Max-Change: 0.27109Iteration: 13, Log-Lik: -11483.832, Max-Change: 0.20093Iteration: 14, Log-Lik: -11478.736, Max-Change: 0.36957Iteration: 15, Log-Lik: -11473.191, Max-Change: 0.30059Iteration: 16, Log-Lik: -11469.110, Max-Change: 0.21768Iteration: 17, Log-Lik: -11467.036, Max-Change: 0.24581Iteration: 18, Log-Lik: -11465.232, Max-Change: 0.13526Iteration: 19, Log-Lik: -11464.412, Max-Change: 1.58684Iteration: 20, Log-Lik: -11462.570, Max-Change: 0.12805Iteration: 21, Log-Lik: -11461.864, Max-Change: 1.49706Iteration: 22, Log-Lik: -11460.761, Max-Change: 0.16832Iteration: 23, Log-Lik: -11460.412, Max-Change: 1.92410Iteration: 24, Log-Lik: -11459.357, Max-Change: 0.10853Iteration: 25, Log-Lik: -11459.336, Max-Change: 4.69566Iteration: 26, Log-Lik: -11458.325, Max-Change: 0.03939Iteration: 27, Log-Lik: -11458.166, Max-Change: 0.99933Iteration: 28, Log-Lik: -11458.050, Max-Change: 0.61374Iteration: 29, Log-Lik: -11457.969, Max-Change: 0.15155Iteration: 30, Log-Lik: -11457.932, Max-Change: 0.00230Iteration: 31, Log-Lik: -11457.931, Max-Change: 0.00265Iteration: 32, Log-Lik: -11457.923, Max-Change: 0.00253Iteration: 33, Log-Lik: -11457.919, Max-Change: 0.00075Iteration: 34, Log-Lik: -11457.918, Max-Change: 0.00064Iteration: 35, Log-Lik: -11457.916, Max-Change: 0.00006