ShinyItemAnalysis: Test and Item Analysis via shiny

The ShinyItemAnalysis package contains an interactive shiny application for the psychometric analysis of educational tests, psychological assessments, health-related and other types of multi-item measurements, or ratings from multiple raters, which can be accessed using function startShinyItemAnalysis(). The shiny application covers a broad range of psychometric methods and offers data examples, model equations, parameter estimates, interpretation of results, together with a selected R code, and is therefore suitable for teaching psychometric concepts with R. It also allows the users to upload and analyze their own data and to automatically generate analysis reports in PDF or HTML.

Besides, the package provides its own functions for test and item analysis within classical test theory framework (e.g., functions gDiscrim(), ItemAnalysis(), DistractorAnalysis(), or DDplot()), using various regression models (e.g., plotCumulative(), plotAdjacent(), plotMultinomial(), or plotDIFLogistic()), and under IRT framework (e.g., ggWrightMap(), or plotDIFirt()).

Package also contains several demonstration datasets including the HCI dataset from the reference paper by Martinkova and Drabinova (2018).

Details

Package: ShinyItemAnalysis
Type: Package
Version: 1.3.8
Date: 2021-09-07
Depends: R (>= 3.5.0)
Imports: cowplot, data.table, deltaPlotR, difNLR (>= 1.3.2), difR (>= 5.0), dplyr, DT, ggdendro, ggplot2, gridExtra, knitr, latticeExtra, lme4, ltm, magrittr, mirt (>= 1.24), msm, nnet, plotly, psych, psychometric, purrr, rlang, rmarkdown, rstudioapi, scales, shiny (>= 1.0.3), shinyBS, shinydashboard, shinyjs (>= 0.9), stringr, tibble, tidyr, VGAM, xtable License: GPL-3
BugReports: https://github.com/patriciamar/ShinyItemAnalysis/issues
Website: http://www.ShinyItemAnalysis.org/
Encoding: UTF-8

Functions

startShinyItemAnalysis

DDplot

DistractorAnalysis

plotDistractorAnalysis

gDiscrim

ggWrightMap

ItemAnalysis

plotAdjacent, plotCumulative, plotMultinomial

plotDIFirt, plotDIFLogistic

plot_corr

recode_nr

ICCrestricted

fa_parallel

Datasets

AIBS

AttitudesExpulsion

CZmatura

CZmaturaS

dataMedical

dataMedicalgraded

dataMedicalkey

dataMedicaltest

HCI

HCIdata

HCIgrads

HCIkey

HCIprepost

HCItest

HCItestretest

LearningToLearn

NIH

References

Martinkova, P., & Drabinova, A. (2018). ShinyItemAnalysis for teaching psychometrics and to enforce routine analysis of educational tests. The R Journal, 10(2), 503--515, doi: 10.32614/RJ-2018-074

See also

Author

Patricia Martinkova
Institute of Computer Science of the Czech Academy of Sciences
Faculty of Education, Charles University
martinkova@cs.cas.cz

Adela Hladka (nee Drabinova)
Institute of Computer Science of the Czech Academy of Sciences

Jan Netik
Institute of Computer Science of the Czech Academy of Sciences
netik@cs.cas.cz