Compute generalized item discrimination

gDiscrim function computes various generalizations of discrimination index ULI. It enumerates the ability of item to distinguish between individuals from upper (U) vs. lower (L) ability groups, i.e. between respondents with high vs. low overall score on the test. Number of groups, as well as upper and lower groups can be specified by user. Maximal and minimal score in ordinal datasets can be specified by user.

gDiscrim(Data, k = 3, l = 1, u = 3, maxscore, minscore, x)

Arguments

Data
matrix or data.frame of items to be examined. Rows represent respondents, columns represent items.

k
numeric: number of groups to which may be Data divided by the total score. Default value is 3. See Details.

l
numeric: lower group. Default value is 1. See Details.

u
numeric: upper group. Default value is 3. See Details.

maxscore
numeric: maximal score in ordinal items. If missing, vector of obtained maximal scores is imputed. See Details.

minscore
numeric: minimal score in ordinal items. If missing, vector of obtained minimal scores is imputed. See Details.

x
deprecated. Use argument Data instead.

Details

The function computes total test scores for all respondents and then divides the respondents into k groups. The lower and upper groups are determined by l and u parameters, i.e., l-th and u-th group where the ordering is defined by increasing total score.

In ordinal items, difficulty is calculated as difference of average score divided by range (maximal possible score maxscore minus minimal possible score minscore for given item).

Discrimination is calculated as difference in difficulty between upper and lower group.

Note

gDiscrim is used by DDplot function.

References

Martinkova, P., Stepanek, L., Drabinova, A., Houdek, J., Vejrazka, M., & Stuka, C. (2017). Semi-real-time analyses of item characteristics for medical school admission tests. In: Proceedings of the 2017 Federated Conference on Computer Science and Information Systems. https://doi.org/10.15439/2017F380

See also

Author

Adela Hladka
Institute of Computer Science of the Czech Academy of Sciences
hladka@cs.cas.cz

Lubomir Stepanek
Charles University

Jana Vorlickova
Institute of Computer Science of the Czech Academy of Sciences

Patricia Martinkova
Institute of Computer Science of the Czech Academy of Sciences
martinkova@cs.cas.cz

Examples

# loading 100-item medical admission test datasets
data(dataMedical, dataMedicalgraded)
# binary dataset
dataBin <- dataMedical[, 1:100]
# ordinal dataset
dataOrd <- dataMedicalgraded[, 1:100]

# ULI for the first 5 items of binary dataset
# compare to psychometric::discrim(dataBin)
gDiscrim(dataBin)[1:5]

#>     X2001     X2002     X2003     X2004     X2005 
#> 0.1656211 0.1066499 0.4943538 0.3889586 0.3312422 

# generalized ULI using 5 groups, compare 4th and 5th for binary dataset
gDiscrim(dataBin, k = 5, l = 4, u = 5)[1:5]

#>     X2001     X2002     X2003     X2004     X2005 
#> 0.1066946 0.1171548 0.2050209 0.2280335 0.1192469 

# ULI for first 5 items for ordinal dataset
gDiscrim(dataOrd)[1:5]

#>     X2001     X2002     X2003     X2004     X2005 
#> 0.0966123 0.1759724 0.3177541 0.2026349 0.1725220 

# generalized ULI using 5 groups, compare 4th and 5th for binary dataset
gDiscrim(dataOrd, k = 5, l = 4, u = 5)[1:5]

#>      X2001      X2002      X2003      X2004      X2005 
#> 0.04079498 0.12343096 0.08158996 0.09361925 0.06014644 

# maximum (4) and minimum (0) score are same for all items
gDiscrim(dataOrd, k = 5, l = 4, u = 5, maxscore = 4, minscore = 0)[1:5]

#>      X2001      X2002      X2003      X2004      X2005 
#> 0.04079498 0.12343096 0.08158996 0.09361925 0.06014644