Get Original Levels from a Fitted BLIS model

Just a simple accessor to original levels and correct key stored in fitted BLIS model.

Usage

get_orig_levels(object)

Arguments

object

object of class BlisClass, model fitted via fit_blis() or blis().

Value

list of the original levels and correct key. Key is stored as an attribute key for every individual item.

See also

Examples

fit <- fit_blis(HCItest[, 1:20], HCIkey)
#> Iteration: 1, Log-Lik: -15023.783, Max-Change: 4.45438Iteration: 2, Log-Lik: -12347.193, Max-Change: 2.04910Iteration: 3, Log-Lik: -11708.838, Max-Change: 0.82435Iteration: 4, Log-Lik: -11596.801, Max-Change: 0.38587Iteration: 5, Log-Lik: -11551.752, Max-Change: 0.38882Iteration: 6, Log-Lik: -11523.110, Max-Change: 0.36396Iteration: 7, Log-Lik: -11503.629, Max-Change: 0.24205Iteration: 8, Log-Lik: -11490.180, Max-Change: 0.22810Iteration: 9, Log-Lik: -11479.823, Max-Change: 0.19168Iteration: 10, Log-Lik: -11472.895, Max-Change: 0.12438Iteration: 11, Log-Lik: -11468.286, Max-Change: 0.09702Iteration: 12, Log-Lik: -11465.172, Max-Change: 0.07258Iteration: 13, Log-Lik: -11460.741, Max-Change: 0.06162Iteration: 14, Log-Lik: -11459.686, Max-Change: 0.04574Iteration: 15, Log-Lik: -11459.051, Max-Change: 0.02648Iteration: 16, Log-Lik: -11458.423, Max-Change: 0.02381Iteration: 17, Log-Lik: -11458.210, Max-Change: 0.01723Iteration: 18, Log-Lik: -11458.069, Max-Change: 0.02161Iteration: 19, Log-Lik: -11457.896, Max-Change: 0.01266Iteration: 20, Log-Lik: -11457.848, Max-Change: 0.00632Iteration: 21, Log-Lik: -11457.822, Max-Change: 0.00513Iteration: 22, Log-Lik: -11457.772, Max-Change: 0.00228Iteration: 23, Log-Lik: -11457.769, Max-Change: 0.00110Iteration: 24, Log-Lik: -11457.767, Max-Change: 0.00080Iteration: 25, Log-Lik: -11457.765, Max-Change: 0.00155Iteration: 26, Log-Lik: -11457.765, Max-Change: 0.00047Iteration: 27, Log-Lik: -11457.764, Max-Change: 0.00036Iteration: 28, Log-Lik: -11457.764, Max-Change: 0.00016Iteration: 29, Log-Lik: -11457.764, Max-Change: 0.00074Iteration: 30, Log-Lik: -11457.764, Max-Change: 0.00080Iteration: 31, Log-Lik: -11457.764, Max-Change: 0.00088Iteration: 32, Log-Lik: -11457.763, Max-Change: 0.00032Iteration: 33, Log-Lik: -11457.763, Max-Change: 0.00015Iteration: 34, Log-Lik: -11457.763, Max-Change: 0.00042Iteration: 35, Log-Lik: -11457.763, Max-Change: 0.00035Iteration: 36, Log-Lik: -11457.763, Max-Change: 0.00017Iteration: 37, Log-Lik: -11457.763, Max-Change: 0.00007
get_orig_levels(fit)
#> $`Item 1`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE FALSE  TRUE
#> 
#> $`Item 2`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1] FALSE  TRUE FALSE
#> 
#> $`Item 3`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE
#> 
#> $`Item 4`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE FALSE  TRUE
#> 
#> $`Item 5`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE  TRUE FALSE FALSE
#> 
#> $`Item 6`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE
#> 
#> $`Item 7`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE
#> 
#> $`Item 8`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE
#> 
#> $`Item 9`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE FALSE  TRUE
#> 
#> $`Item 10`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE
#> 
#> $`Item 11`
#> [1] "A" "B" "C"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE
#> 
#> $`Item 12`
#> [1] "A" "B" "C" "D" "E"
#> attr(,"key")
#> [1] FALSE FALSE FALSE  TRUE FALSE
#> 
#> $`Item 13`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE FALSE
#> 
#> $`Item 14`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE FALSE
#> 
#> $`Item 15`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE FALSE
#> 
#> $`Item 16`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1]  TRUE FALSE FALSE FALSE
#> 
#> $`Item 17`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE FALSE
#> 
#> $`Item 18`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE FALSE
#> 
#> $`Item 19`
#> [1] "A" "B" "C" "D" "E"
#> attr(,"key")
#> [1] FALSE FALSE  TRUE FALSE FALSE
#> 
#> $`Item 20`
#> [1] "A" "B" "C" "D"
#> attr(,"key")
#> [1] FALSE FALSE FALSE  TRUE
#>